+++ crates/swc_ecma_transforms_module/tests/fixture/common/issue-8047/output.amd.js
@@ -1,0 +1,11 @@
+define([
+    "require"
+], function(require) {
+    "use client";
+    "foo";
+    "use bar";
+    "use strict";
+    // All above are directives
+    function foo() {}
+    "use hello"; // This is not directive
+});